+++ etc/blocks/pool-explorer/ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @milaboratories/milaboratories.pool-explorer.ui
 
+## 1.1.60
+
+### Patch Changes
+
+- Updated dependencies [621e15a]
+  - @platforma-sdk/ui-vue@1.29.12
+  - @milaboratories/milaboratories.pool-explorer.model@1.0.28
+
+## 1.1.59
+
+### Patch Changes
+
+- Updated dependencies [7cd6249]
+  - @platforma-sdk/ui-vue@1.29.9
+
 ## 1.1.58
 
 ### Patch Changes