+++ etc/blocks/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.3.64
+
+### Patch Changes
+
+- Updated dependencies [621e15a]
+  - @platforma-sdk/model@1.29.12
+  - @milaboratories/milaboratories.ui-examples.model@1.1.44
+
 ## 1.3.63
 
 ### Patch Changes