+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/pl-drivers
 
+## 1.5.45
+
+### Patch Changes
+
+- Updated dependencies [621e15a]
+  - @milaboratories/pl-model-common@1.13.6
+
+## 1.5.44
+
+### Patch Changes
+
+- @milaboratories/pl-tree@1.5.5
+
 ## 1.5.43
 
 ### Patch Changes